WIP
I think I've zoned in on the main issue occurring here. In the webhook,
contentful only sends defined field values. Therefore, if a field, is
nullified, it doesn't get included in the contenful request body. When
comparing fields with Algolia, then, it doesn't get removed/reset.

A few debugging ideas:
- Create a reset endpoint by recordID. This would be a quick way for a
developer to reset the algolia records for indivual page details without
having to reset the entire index. Future improvements to this could
include a dashboard or other enhanced tooling
- Improve the webhook post endpoint to make sure the known values we
want to save to algolia are compared. I.e., even if they aren't in the
request body, we know to expect them and adjust accordingly if they are
not.

Both of those ideas are WIP and in this commit.